requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.8.0
pydantic[email] ==2.8.0
pytest==7.4.3
httpx==0.25.1
requests==2.31.0
coverage==6.5.0
pytest-cov==4.1.0

